+++ src/modules/voxel/benchmarks/SurfaceExtractorBenchmark.cpp
@@ -3,7 +3,6 @@
  */
 
 #include "app/benchmark/AbstractBenchmark.h"
-#include "core/collection/DynamicArray.h"
 #include "palette/Palette.h"
 #include "voxel/ChunkMesh.h"
 #include "voxel/RawVolume.h"
@@ -111,18 +110,6 @@ BENCHMARK_DEFINE_F(SurfaceExtractorBenchmark, Cubic)(benchmark::State &state) {
 	}
 }
 
-namespace voxel {
-void prepareChunk(const voxel::RawVolume &map, core::DynamicArray<voxel::Voxel> &voxels, const glm::ivec3 &chunkPos);
-}
-
-BENCHMARK_DEFINE_F(SurfaceExtractorBenchmark, BinaryPrepareChunk)(benchmark::State &state) {
-	for (auto _ : state) {
-		core::DynamicArray<voxel::Voxel> voxels;
-		glm::ivec3 chunkPos = v.region().getLowerCorner() - 1;
-		voxel::prepareChunk(v, voxels, chunkPos);
-	}
-}
-
 BENCHMARK_DEFINE_F(SurfaceExtractorBenchmark, Binary)(benchmark::State &state) {
 	for (auto _ : state) {
 		const bool ambientOcclusion = false;
@@ -149,7 +136,6 @@ BENCHMARK_DEFINE_F(SurfaceExtractorBenchmark, MarchingCubes)(benchmark::State &s
 
 BENCHMARK_REGISTER_F(SurfaceExtractorBenchmark, Cubic);
 BENCHMARK_REGISTER_F(SurfaceExtractorBenchmark, Binary);
-BENCHMARK_REGISTER_F(SurfaceExtractorBenchmark, BinaryPrepareChunk);
 BENCHMARK_REGISTER_F(SurfaceExtractorBenchmark, MarchingCubes);
 
 BENCHMARK_MAIN();